% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerDynamics.R
\name{getDynamics}
\alias{getDynamics}
\title{Clusters the active enhancer into condition-specific groups}
\usage{
getDynamics(data, w_0 = 0.5, cutoff = 0.05, W = 10, C = 1)
}
\arguments{
\item{data}{List containing the metadata of the experiments and
the bin-wise enhancer prediction values as a GRanges object 
(calculated in the prior step)}

\item{w_0}{The minimum difference between the normalized prediction means
that two enhancers need in order to be included in the clustering. 
Default is 0.5 ([0,1]).}

\item{cutoff}{cutoff for the p-values calculated during the clustering.
Default is 0.05 ([0,1]).}

\item{W}{Number of bins +/- the current bin that should be included 
when calculating the p-values. Default is 10 ([5, 50]).}

\item{C}{Number of cores to use}
}
\value{
A list containing the meta data of the experiments and
the condition-specific clusters in a GRanges object
}
\description{
Clusters the active enhancer into condition-specific groups
}
\examples{
                                           
#recreate the outputs of getEnhancers()
files <- c(system.file("extdata", "Condition1.H3K4me1.bam", package="crupR"),
           system.file("extdata", "Condition1.H3K4me3.bam", package="crupR"),
           system.file("extdata", "Condition1.H3K27ac.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K4me1.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K4me3.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K27ac.bam", package="crupR"))
inputs <- rep(system.file("extdata", "Condition1.Input.bam", package="crupR"), 3)
inputs2 <- rep(system.file("extdata", "Condition2.Input.bam", package="crupR"), 3)  
metaData <- data.frame(HM = rep(c("H3K4me1", "H3K4me3", "H3K27ac"),2),
                       condition = c(1,1,1,2,2,2), replicate = c(1,1,1,1,1,1),
                       bamFile = files, inputFile = c(inputs, inputs2))
metaData1 <- subset(metaData, condition == 1)
metaData2 <- subset(metaData, condition == 2)
pred1 <- readRDS(system.file("extdata", "condition1_predictions.rds", package = "crupR"))
pred1 <- list("metaData" = metaData1, "D" = pred1)
pred2 <- readRDS(system.file("extdata", "condition2_predictions.rds", package = "crupR"))
pred2 <- list("metaData" = metaData2, "D" = pred2)
#put the outputs in a list
predictions <- list(pred1, pred2)
#run the function
getDynamics(data = predictions, C = 2)

}
