% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerPrediction.R
\name{getEnhancers}
\alias{getEnhancers}
\title{Predicts the occurence of enhancers using the normalized HMs counts.}
\usage{
getEnhancers(data, classifier = NULL, all = FALSE, C = 1)
}
\arguments{
\item{data}{List containing the metaData of the experiments that were normalized and also the normalized counts as GRanges object (basically the output of the normalize() step)}

\item{classifier}{The path of the classifier to use for the prediction. When set to NULL (default), the default clasifier is used.}

\item{all}{Output of separated probabilities. Default is FALSE.}

\item{C}{Number of cores to use. Default is 1.}
}
\value{
A list containing the meta data of the experiments whose enhancers were predicted and
the enhancer probabilities for each 100 bp bin in the genome as a GRanges object
}
\description{
Predicts the occurence of enhancers using the normalized HMs counts.
}
\examples{
#first recreate the output of crupR::normalize (so skip this)
files <- c(system.file("extdata", "Condition2.H3K4me1.bam", package="crupR"),
          system.file("extdata", "Condition2.H3K4me3.bam", package="crupR"),
          system.file("extdata", "Condition2.H3K27ac.bam", package="crupR"))
inputs <- rep(system.file("extdata", "Condition2.Input.bam", package="crupR")) 
metaData <- data.frame(HM = c("H3K4me1","H3K4me3","H3K27ac"),
                    condition = c(2,2,2), replicate = c(1,1,1),
                    bamFile = files, inputFile = inputs)
d <- readRDS(system.file("extdata", "condition2_normalized.rds", package="crupR"))
norm <- list(metaData = metaData, D = d)
#let's run the actual function
getEnhancers(data = norm, C = 2)

}
