% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_superEnhancers.R
\name{getSE}
\alias{getSE}
\title{Uses the enhancer predictions to find clusters of enhancers ('superenhancer' - like).}
\usage{
getSE(data, cutoff = 0.5, distance = 12500, C = 1)
}
\arguments{
\item{data}{List containing the meta data of the experiments and the enhancer prediction values in a GRanges object calculated by enhancerPrediction()}

\item{cutoff}{Cutoff for enhancer probabilities. Default is 0.5.}

\item{distance}{Maximum distance (bp) for clustering. Default is 12500}

\item{C}{Number of cores to use}
}
\value{
A list containing the meta data of the experiments, the enhancer prediction values as a GRanges object (like the input data),
the enhancer peak calls as a GRanges object (can be exported as a bedGraph) and the cluster of peaks in a GRanges object (can be exported as bed file)
}
\description{
Uses the enhancer predictions to find clusters of enhancers ('superenhancer' - like).
}
\examples{

# first recreate the output of crupR:getPredictions
files <- c(system.file("extdata", "Condition2.H3K4me1.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K4me3.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K27ac.bam", package="crupR"))
inputs <- rep(system.file("extdata", "Condition2.Input.bam", package="crupR"), 3)
#create the metaData frame
metaData <- data.frame(HM = c("H3K4me1","H3K4me3","H3K27ac"),
                       condition = c(2,2,2), replicate = c(1,1,1),
                       bamFile = files, inputFile = inputs)
pred <- readRDS(system.file("extdata", "condition2_predictions.rds", package="crupR"))                                                                                           
prediction <- list(metaData = metaData, D = pred)
#run the function
getSE(data = prediction, C = 2)
}
