% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerTargets.R
\name{getTargets}
\alias{getTargets}
\title{Find the target genes of the regulatory, condition-specific clusters}
\usage{
getTargets(
  data,
  expr,
  genome,
  TAD.file = NULL,
  cutoff = 0.9,
  nearest = FALSE,
  C = 1
)
}
\arguments{
\item{data}{Outputfile of the previous step (getDynamics())}

\item{expr}{GRanges object containing the gene expression counts of RNAseq experiments for each condition and its replicates}

\item{genome}{Genome used in the .bam files of the RNAseq experiments. Possible options are 'mm9', 'mm10', 'hg19' and 'hg38'.}

\item{TAD.file}{Path to the TAD file to use for finding the target genes. If set to NULL, the default file is used (only if the 'mm10' genome was used)}

\item{cutoff}{cutoff for correlation between cluster and gene. Default is 0.9.}

\item{nearest}{Logical: if set, the nearest gene is taken to build the regulatory regions.}

\item{C}{Number of cores to use}
}
\value{
A list containing the meta data of the experiments
and a GRanges object containing the dynamic regulatory units
}
\description{
Find the target genes of the regulatory, condition-specific clusters
}
\examples{

#first get the output of crupR::getDynamics so skip this
files <- c(system.file("extdata", "Condition1.H3K4me1.bam", package="crupR"),
           system.file("extdata", "Condition1.H3K4me3.bam", package="crupR"),
           system.file("extdata", "Condition1.H3K27ac.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K4me1.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K4me3.bam", package="crupR"),
           system.file("extdata", "Condition2.H3K27ac.bam", package="crupR"))
inputs <- rep(system.file("extdata", "Condition1.Input.bam", package="crupR"), 3)
inputs2 <- rep(system.file("extdata", "Condition2.Input.bam", package="crupR"), 3)  
metaData <- data.frame(HM = rep(c("H3K4me1", "H3K4me3", "H3K27ac"),2),
                       condition = c(1,1,1,2,2,2), replicate = c(1,1,1,1,1,1),
                       bamFile = files, inputFile = c(inputs, inputs2))
clusters <- readRDS(system.file("extdata", "differential_enhancers.rds", package="crupR"))
dynamics <- list("metaData" = metaData, "sumFile" = clusters)
#load your GRanges object containing the gene expressions counts and run the function
expr <- readRDS(system.file("extdata", "expressions.rds", package="crupR"))
getTargets(data = dynamics, expr = expr, genome = "mm10", C = 2)

}
