% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSummary}
\alias{plotSummary}
\title{plots boxplots of the median enhancer prediction values of the enhancers in the condition-specific clusters}
\usage{
plotSummary(dynamic_enhancers, num_plots = 20)
}
\arguments{
\item{dynamic_enhancers}{The getDynamics() output file of containing the GRanges object with the differential enhaners}
}
\value{
a figure containing a boxplot for each cluster
}
\description{
plots boxplots of the median enhancer prediction values of the enhancers in the condition-specific clusters
}
\examples{

#get the output of crupR::getTargets
dynamics <-readRDS(system.file("extdata", "differential_enhancers.rds", package="crupR"))
plotSummary(dynamics)#plot the clusters

}
